Spike out a minimal default system with Prism
This change implements the minimum of what we could do by using Prism to
parse out the default to create a zero-cost initializer with default
values.

While it works, it's an MVP. I suspect that if you were to split the
default lambda into multiple lines, it would choke. I know that you
can't use it in REPLs since you don't have a source location that makes
sense for parsing.

But it's a start at an idea for what we could do.