Support non single-line default procs
This change fleshes out the Prism parsing more so that defaults don't
have to be single-line stabby lambdas. They can't be _any_ expression
and still need to be in a lambda, but can span multiple lines.

I'm sure I have missed corner cases here, but we're to the point where I
can say the general idea works. The compile performance is
abysmal (roughly 500x slower than writing a PORO), but I think that's a
worthwhile trade-off. It's also rife with opportunities for optimization
as it's still a proof of concept. It currently parses the entire file N
times where N is the number of consts/props in the file.

This work should be doable in a single parsing pass where we cache the
default of every const/prop in a file. After that, we could have a hook
that clears that cache so there is no runtime memory hit.