Reorganize prop classes
This change removes the props classes from the Struct class so that they
don't pollute any subclasses. Now, the only public methods exposed on
the Struct are `define` and `props` and the only public methods on Props
are `const`, `prop`, and `call` (which is incidentally public so that
the Struct can call it).

This make it less likely that we'll conflict with subclass methods and
improves communication of the API.