Add Sorbet setup
This change adds the setup for Sorbet and Tapioca. The reason for this
is that this gem is intended to work well with Sorbet and in order to do
that, I feel that we need to have Sorbet working well within the
repository.

I'm not wild at how much Sorbet bloats the repository, but I'd rather
have the RBI for the gem ship with the gem instead of shipping it as an
annotation from a repository.