Simplify tracking the define block
This change uses the caller of the define block explicitly to remove the
need to find the struct as part of extracting defaults. By removing a
whole layer of the onion, there are now fewer moving parts and it's
easier to reason about.

This does not give a performance boost, but it doesn't cost us
significantly in performance either; the difference falls within a
p=0.05 margin of error.